# performance improvements when encoding large arrays and maps