--- a/src/Schema.ts
+++ b/src/Schema.ts
@@ -628,8 +628,9 @@ export abstract class Schema {
                 if (!encodeAll) {
                     value.$changes.discard();
 
-                    // TODO: track array/map indexes per client?
+                    // TODO: track array/map indexes per client (for filtering)?
                     if (!client) {
+                        // TODO: do not iterate though all MapSchema indexes here.
                         this[`_${field}`]._updateIndexes();
                     }
                 }