--- a/test/PerformanceTest.ts
+++ b/test/PerformanceTest.ts
@@ -9,8 +9,7 @@ function assertExecutionTime(cb: Function, message: string, threshold: number) {
     cb();
     const diff = Date.now() - now;
     console.log(`${message} took ${diff}ms`)
-    assert.ok(diff < threshold, `${message} exceeded ${threshold}ms. took: ${diff}ms`);
-
+    assert.ok(diff <= threshold, `${message} exceeded ${threshold}ms. took: ${diff}ms`);
 }
 
 describe("Performance", () => {
@@ -26,7 +25,7 @@ describe("Performance", () => {
             }
         }, `inserting ${totalItems} items to array`, 1200); // TODO: improve this!
 
-        assertExecutionTime(() => state.encode(), `encoding ${totalItems} array entries`, 150);
+        assertExecutionTime(() => state.encode(), `encoding ${totalItems} array entries`, 190);
 
         const player: Player = state.arrayOfPlayers[Math.round(totalItems / 2)];
         player.x = getRandomNumber();
@@ -35,19 +34,21 @@ describe("Performance", () => {
         assertExecutionTime(() => state.encode(), "encoding a single array item change", 5);
     });
 
-    it("MapSchema", () => {
+    it("MapSchema", function (){
+        this.timeout(10000);
+
         const state = new State();
         state.mapOfPlayers = new MapSchema<Player>();
 
-        const totalItems = 5000;
+        const totalItems = 10000;
 
         assertExecutionTime(() => {
             for (let i = 0; i < totalItems; i++) {
                 state.mapOfPlayers["player" + i] = new Player("Player " + i, getRandomNumber(), getRandomNumber());
             }
-        }, `inserting ${totalItems} items to map`, 600); // TODO: improve this value!
+        }, `inserting ${totalItems} items to map`, 2600); // TODO: improve this value!
 
-        assertExecutionTime(() => state.encode(), `encoding ${totalItems} map entries`, 100);
+        assertExecutionTime(() => state.encode(), `encoding ${totalItems} map entries`, 150);
 
         const player: Player = state.mapOfPlayers[`player${Math.floor(totalItems / 2)}`];
         player.x = getRandomNumber();