I've replaced `bytes = bytes.concat(moreBytes)` with `bytes.push(...moreBytes)` and added a `PerformanceTest` to the test-suite to assert the amount of time to encode large arrays and maps.

**What the test file checks?**

- Time spent adding 10000 items to a map/array
- Time spent encoding 10000 items of a map/array
- Time spent encoding a single item update of the map/array

## A few observations/TODO's

### `ArraySchema`

- [ ] Adding all values to `ArraySchema` is costly (takes ~790ms)
- [x] Encoding `ArraySchema` is reasonable (takes ~116ms)
- [x] Encoding a single item change on `ArraySchema` is perfect (takes ~0ms)

### `MapSchema`
- [ ] Adding all values to `MapSchema` is costly (takes ~1416ms)
- [x] Encoding `MapSchema` is reasonable (takes ~91ms)
- [ ] Encoding a single item change on `MapSchema` needs improvement (takes ~4ms)