@endel 2019-09-27
Some more work is needed for `MapSchema`, as the size of maps increase, the time for encoding is increasing as well, because of this call: https://github.com/colyseus/schema/blob/4f97a02446cdef9d5343a31a9d3d7bdcddd6b05a/src/Schema.ts#L633-L634